Move creds forms to shared
