+++ shared/components/CredentialForms/BasicCredentialsForm.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { CustomPropTypes } from 'react-shared';
-import TextFormItem from './../../../Shared/TextFormItem';
+import CustomPropTypes from '../../typechecking/CustomPropTypes';
+import { TextFormItem } from '../TextFormItem/TextFormItem';
 
 export const CREDENTIAL_TYPE_BASIC = 'Basic';
 
@@ -15,7 +15,7 @@ BasicCredentialsForm.propTypes = {
   defaultValues: PropTypes.object,
 };
 
-export default function BasicCredentialsForm({ refs, defaultValues }) {
+export function BasicCredentialsForm({ refs, defaultValues }) {
   return (
     <section className="fd-has-margin-top-medium">
       <TextFormItem
