+++ shared/components/CredentialForms/index.js
@@ -1,0 +1,3 @@
+export * from './CredentialsForm';
+export * from './BasicCredentialsForm';
+export * from './OAuthCredentialsForm';
